+++ Cargo.toml
@@ -1,3 +1,6 @@
 [workspace]
 resolver = "2"
 members = [ "crates/cli", "crates/common","crates/server", "crates/tui"]
+
+[workspace.dependencies]
+common = {path = "crates/common", version = "0.1.0"}
