+++ crates/common/Cargo.toml
@@ -4,3 +4,5 @@ version = "0.1.0"
 edition = "2024"
 
 [dependencies]
+reqwest = { version = "0.12", features = ["json"] }
+serde = { version = "1.0.192", features = ["derive"] }
